fix: DatePillsBar no longer eagerly fetch entries
Previously the eager fetching of displayed dates are creating duplicate
entry fetches because the form are also requesting its own data fetch.
This fixes a lot of the race condition issues at the cost of having a
tiny flash of unloaded entry if user clicks on an entry for previous
month.