feat: submitting entries + rearchitecture
In the original design, the browser would connect directly to the backend legacy API for performance.
This works for fetching entries with GET request, but causes problems when it comes to submitting entries due to CORS.

To resolve this, I made the decision to redesign the system to have the browser connect only interact with SvelteKit backend, even when using the legacy API. This will cause an additional hop, but has an additional benefit of simplifying front-end architecture to not have to talk to 2 different APIs.

For now, the entry fetching are still done directly via browser, but the submission is now done via SvelteKit as middleman.